Change license conditions to a variant of the BSD license
The warranty disclaimer in the zlib license may be too weak.  We now use
a modified BSD license that (i) requires to retain the license text when
distributing the packages in source form but allows to omit this text
when distributing the packages in binary form, and (ii) requires to
document changes made to the files when distributing modified source
versions.