emacs: Start Emacs from a shell script
Emacs requires the LC_NUMERIC locale to be set to "C" because its Lisp
parser is locale-dependent.  Emacs sets this locale during initialization,
but GTK+ and GStreamer reset it to the one specified by the environment.
To avoid problems with locales that use a decimal separator other than
".", we now start Emacs from a shell script that sets the LC_NUMERIC
environment variable to "C" before running Emacs.

(fixes bnc#779426)